Entity Changes for ivory

+++ src/main/java/com/maan/eway/bean/MsVehicleDetails.java
@@ -16,6 +16,8 @@ package com.maan.eway.bean;
 import org.hibernate.annotations.DynamicInsert;
 import org.hibernate.annotations.DynamicUpdate;
 
+import com.fasterxml.jackson.annotation.JsonProperty;
+
 import lombok.*;
 import javax.persistence.*;
 import java.io.Serializable;
@@ -325,10 +327,11 @@ private static final long serialVersionUID = 1L;
 	private Double lossRatio;
 	
 	@Column(name = "HORSE_POWER")
-	private Double horsePower;
+	private Integer horsePower;
 	
-	@Column(name = "ZONE")
-	private Double zone;
+	  @Column(name = "ZONE")
+		private Integer zone;
+	
 }
